requirements.txt
asgiref==3.7.2
click==8.1.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
joblib==1.3.1
nltk==3.8.1
numpy==1.25.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.11.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.65.0
typing_extensions==4.7.1
